outsource VM types to prevent circular imports
